--- library/hook/src/main/java/com/sevtinge/hyperceiler/hook/module/hook/systemframework/ToastBlur.java
@@ -1,82 +1,0 @@
-/*
-  * This file is part of HyperCeiler.
-
-  * HyperCeiler is free software: you can redistribute it and/or modify
-  * it under the terms of the GNU Affero General Public License as
-  * published by the Free Software Foundation, either version 3 of the
-  * License.
-
-  * This program is distributed in the hope that it will be useful,
-  * but WITHOUT ANY WARRANTY; without even the implied warranty of
-  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  * GNU Affero General Public License for more details.
-
-  * You should have received a copy of the GNU Affero General Public License
-  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-  * Copyright (C) 2023-2025 HyperCeiler Contributions
-*/
-package com.sevtinge.hyperceiler.hook.module.hook.systemframework;
-
-import static com.sevtinge.hyperceiler.hook.module.base.tool.HookTool.findAndHookMethod;
-import static de.robv.android.xposed.XposedHelpers.findClassIfExists;
-
-import android.content.Context;
-import android.view.View;
-
-import com.sevtinge.hyperceiler.hook.module.base.BaseHook;
-
-import de.robv.android.xposed.XC_MethodHook;
-import de.robv.android.xposed.XposedHelpers;
-
-public class ToastBlur extends BaseHook {
-    private static final String TAG = "ToastBlur";
-
-    @Override
-    public void init() {
-        /*findAndHookMethod("android.widget.ToastStubImpl", classLoader, "addBlur",
-                View.class, Context.class, boolean.class,
-                new XC_MethodHook() {
-                    @Override
-                    protected void beforeHookedMethod(MethodHookParam param) {
-                        // logE(TAG, "view: " + param.args[0] + " con: " + param.args[1] + " boo: " + param.args[2]);
-                    }
-                }
-        );*/
-
-        findAndHookMethod("android.widget.ToastPresenter", lpparam.classLoader, "trySendAccessibilityEvent",
-                View.class, String.class,
-                new XC_MethodHook() {
-                    @Override
-                    protected void beforeHookedMethod(MethodHookParam param) {
-                        boolean mCustomizeView = false;
-                        View mView = (View) XposedHelpers.getObjectField(param.thisObject, "mView");
-                        if (mView == null) return;
-                        if (mView != null) {
-                            mCustomizeView = true;
-                        }
-                        Object toastStub;
-                        Class<?> ToastStub = findClassIfExists("android.widget.ToastStub", lpparam.classLoader);
-                        Context mContext = mView.getContext();
-                        if (ToastStub != null) {
-                            toastStub = XposedHelpers.callStaticMethod(ToastStub, "get");
-                            XposedHelpers.callMethod(toastStub, "addBlur", mView, mContext, false);
-                        }
-                    }
-                }
-        );
-
-        /*findAndHookMethod("android.widget.ToastStubImpl$1", classLoader,
-                "onLayoutChange",
-                View.class, int.class, int.class, int.class,
-                int.class, int.class, int.class, int.class, int.class,
-                new XC_MethodHook() {
-                    @Override
-                    protected void beforeHookedMethod(MethodHookParam param) {
-                        // logE(TAG, "change");
-                    }
-                }
-        );*/
-    }
-
-}
